refactor(GameService): service 패키지 생성 후 파일 트리 구조 리팩토링
