+++ docs/README.md
@@ -45,15 +45,15 @@
   - System.exit()를 호출하지 않는다.
 
 # 추가된 요구 사항
-- [ ] indent(인덴트, 들여쓰기) depth를 3이 넘지 않도록 구현한다. 2까지만 허용한다. 
+- [x] indent(인덴트, 들여쓰기) depth를 3이 넘지 않도록 구현한다. 2까지만 허용한다. 
   - 예를 들어 while문 안에 if문이 있으면 들여쓰기는 2이다. 
   - 힌트: indent(인덴트, 들여쓰기) depth를 줄이는 좋은 방법은 함수(또는 메서드)를 분리하면 된다.  
   
  
-- [ ] 3항 연산자를 쓰지 않는다.  
+- [x] 3항 연산자를 쓰지 않는다.  
   
 
-- [ ] 함수(또는 메서드)가 한 가지 일만 하도록 최대한 작게 만들어라.  
+- [x] 함수(또는 메서드)가 한 가지 일만 하도록 최대한 작게 만들어라.  
   
 
 - [ ] JUnit 5와 AssertJ를 이용하여 본인이 정리한 기능 목록이 정상 동작함을 테스트 코드로 확인한다.
